requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# FastAPI and server
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# OpenAI integration (for original agent_a.py)
openai>=1.3.7
